Refactor: properties and exception
