+++ src/main/java/mejai/mejaigg/riot/service/RiotProperties.java
@@ -2,12 +2,12 @@ package mejai.mejaigg.riot.service;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
-import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.Setter;
 
 @Getter
-@ConfigurationProperties(prefix = "riot.api")
-@AllArgsConstructor
+@Setter
+@ConfigurationProperties(prefix = "riot")
 public class RiotProperties {
 	private String apiKey;
 }
